Rename _BYTES constants to _MAX_BYTES when appropriate.
None of them is the exact size of their respective serialized data
structure. None of them is the minimum size of their respective
serialized data structure.

This is because DH_MPI_MAX_BYTES is the maximum size for a DH MPI, that
is 4 + (0...384), since a MPI has no leading zeroes and is mod P.