Remove redundant wrapper functions for otrng_bool
The otrng_bool is compatible with the C standard boolean logic so the
`otrng_bool_is_true` wrapper isn't needed. The conversion function
`c_bool_to_otrng_bool` is unused across the code base and can also be
safely removed.

The intent of this change is to declutter the code a bit. For
`otrng_bool_is_true` to serve as an abstraction, one would have to
apply it consistently wherever otrng_bool is examined. That would
clutter the code for little practical value though so it seems better
to just remove it.

It's possible of course that I misunderstand the purpose of
`otrng_bool`, and perhaps there are future plans conflicting with this
change. If so I apologise for the noise! :)